+++ scripts/rollback/rollback-3.10.7.sh
@@ -1,0 +1,286 @@
+#!/bin/bash
+# timestamp: 1675785311396
+
+npm dist-tag add @midwayjs/fc-starter@3.10.6 latest
+npm dist-tag add @midwayjs/fc-starter@3.10.7 beta
+npm dist-tag add @midwayjs/async-hooks-context-manager@3.10.6 latest
+npm dist-tag add @midwayjs/async-hooks-context-manager@3.10.7 beta
+npm dist-tag add @midwayjs/axios@3.10.6 latest
+npm dist-tag add @midwayjs/axios@3.10.7 beta
+npm dist-tag add @midwayjs/bootstrap@3.10.6 latest
+npm dist-tag add @midwayjs/bootstrap@3.10.7 beta
+npm dist-tag add @midwayjs/bull-board@3.10.6 latest
+npm dist-tag add @midwayjs/bull-board@3.10.7 beta
+npm dist-tag add @midwayjs/bull@3.10.6 latest
+npm dist-tag add @midwayjs/bull@3.10.7 beta
+npm dist-tag add @midwayjs/cache@3.10.6 latest
+npm dist-tag add @midwayjs/cache@3.10.7 beta
+npm dist-tag add @midwayjs/captcha@3.10.6 latest
+npm dist-tag add @midwayjs/captcha@3.10.7 beta
+npm dist-tag add @midwayjs/casbin-redis-adapter@3.10.6 latest
+npm dist-tag add @midwayjs/casbin-redis-adapter@3.10.7 beta
+npm dist-tag add @midwayjs/casbin-typeorm-adapter@3.10.6 latest
+npm dist-tag add @midwayjs/casbin-typeorm-adapter@3.10.7 beta
+npm dist-tag add @midwayjs/casbin@3.10.6 latest
+npm dist-tag add @midwayjs/casbin@3.10.7 beta
+npm dist-tag add @midwayjs/code-dye@3.10.6 latest
+npm dist-tag add @midwayjs/code-dye@3.10.7 beta
+npm dist-tag add @midwayjs/consul@3.10.6 latest
+npm dist-tag add @midwayjs/consul@3.10.7 beta
+npm dist-tag add @midwayjs/core@3.10.6 latest
+npm dist-tag add @midwayjs/core@3.10.7 beta
+npm dist-tag add @midwayjs/cos@3.10.6 latest
+npm dist-tag add @midwayjs/cos@3.10.7 beta
+npm dist-tag add @midwayjs/cross-domain@3.10.6 latest
+npm dist-tag add @midwayjs/cross-domain@3.10.7 beta
+npm dist-tag add @midwayjs/decorator@3.10.6 latest
+npm dist-tag add @midwayjs/decorator@3.10.7 beta
+npm dist-tag add @midwayjs/etcd@3.10.6 latest
+npm dist-tag add @midwayjs/etcd@3.10.7 beta
+npm dist-tag add @midwayjs/express-session@3.10.6 latest
+npm dist-tag add @midwayjs/express-session@3.10.7 beta
+npm dist-tag add @midwayjs/faas@3.10.6 latest
+npm dist-tag add @midwayjs/faas@3.10.7 beta
+npm dist-tag add @midwayjs/grpc@3.10.6 latest
+npm dist-tag add @midwayjs/grpc@3.10.7 beta
+npm dist-tag add @midwayjs/http-proxy@3.10.6 latest
+npm dist-tag add @midwayjs/http-proxy@3.10.7 beta
+npm dist-tag add @midwayjs/i18n@3.10.6 latest
+npm dist-tag add @midwayjs/i18n@3.10.7 beta
+npm dist-tag add @midwayjs/info@3.10.6 latest
+npm dist-tag add @midwayjs/info@3.10.7 beta
+npm dist-tag add @midwayjs/jwt@3.10.6 latest
+npm dist-tag add @midwayjs/jwt@3.10.7 beta
+npm dist-tag add @midwayjs/kafka@3.10.6 latest
+npm dist-tag add @midwayjs/kafka@3.10.7 beta
+npm dist-tag add @midwayjs/mikro@3.10.6 latest
+npm dist-tag add @midwayjs/mikro@3.10.7 beta
+npm dist-tag add @midwayjs/mock@3.10.6 latest
+npm dist-tag add @midwayjs/mock@3.10.7 beta
+npm dist-tag add @midwayjs/mongoose@3.10.6 latest
+npm dist-tag add @midwayjs/mongoose@3.10.7 beta
+npm dist-tag add @midwayjs/oss@3.10.6 latest
+npm dist-tag add @midwayjs/oss@3.10.7 beta
+npm dist-tag add @midwayjs/otel@3.10.6 latest
+npm dist-tag add @midwayjs/otel@3.10.7 beta
+npm dist-tag add @midwayjs/passport@3.10.6 latest
+npm dist-tag add @midwayjs/passport@3.10.7 beta
+npm dist-tag add @midwayjs/process-agent@3.10.6 latest
+npm dist-tag add @midwayjs/process-agent@3.10.7 beta
+npm dist-tag add @midwayjs/prometheus-socket-io@3.10.6 latest
+npm dist-tag add @midwayjs/prometheus-socket-io@3.10.7 beta
+npm dist-tag add @midwayjs/prometheus@3.10.6 latest
+npm dist-tag add @midwayjs/prometheus@3.10.7 beta
+npm dist-tag add @midwayjs/rabbitmq@3.10.6 latest
+npm dist-tag add @midwayjs/rabbitmq@3.10.7 beta
+npm dist-tag add @midwayjs/redis@3.10.6 latest
+npm dist-tag add @midwayjs/redis@3.10.7 beta
+npm dist-tag add @midwayjs/security@3.10.6 latest
+npm dist-tag add @midwayjs/security@3.10.7 beta
+npm dist-tag add @midwayjs/sequelize@3.10.6 latest
+npm dist-tag add @midwayjs/sequelize@3.10.7 beta
+npm dist-tag add @midwayjs/session@3.10.6 latest
+npm dist-tag add @midwayjs/session@3.10.7 beta
+npm dist-tag add @midwayjs/socketio@3.10.6 latest
+npm dist-tag add @midwayjs/socketio@3.10.7 beta
+npm dist-tag add @midwayjs/static-file@3.10.6 latest
+npm dist-tag add @midwayjs/static-file@3.10.7 beta
+npm dist-tag add @midwayjs/swagger@3.10.6 latest
+npm dist-tag add @midwayjs/swagger@3.10.7 beta
+npm dist-tag add @midwayjs/tablestore@3.10.6 latest
+npm dist-tag add @midwayjs/tablestore@3.10.7 beta
+npm dist-tag add @midwayjs/typegoose@3.10.6 latest
+npm dist-tag add @midwayjs/typegoose@3.10.7 beta
+npm dist-tag add @midwayjs/typeorm@3.10.6 latest
+npm dist-tag add @midwayjs/typeorm@3.10.7 beta
+npm dist-tag add @midwayjs/upload@3.10.6 latest
+npm dist-tag add @midwayjs/upload@3.10.7 beta
+npm dist-tag add @midwayjs/validate@3.10.6 latest
+npm dist-tag add @midwayjs/validate@3.10.7 beta
+npm dist-tag add @midwayjs/version@3.10.6 latest
+npm dist-tag add @midwayjs/version@3.10.7 beta
+npm dist-tag add @midwayjs/view-ejs@3.10.6 latest
+npm dist-tag add @midwayjs/view-ejs@3.10.7 beta
+npm dist-tag add @midwayjs/view-nunjucks@3.10.6 latest
+npm dist-tag add @midwayjs/view-nunjucks@3.10.7 beta
+npm dist-tag add @midwayjs/view@3.10.6 latest
+npm dist-tag add @midwayjs/view@3.10.7 beta
+npm dist-tag add @midwayjs/express@3.10.6 latest
+npm dist-tag add @midwayjs/express@3.10.7 beta
+npm dist-tag add @midwayjs/koa@3.10.6 latest
+npm dist-tag add @midwayjs/koa@3.10.7 beta
+npm dist-tag add @midwayjs/web@3.10.6 latest
+npm dist-tag add @midwayjs/web@3.10.7 beta
+npm dist-tag add @midwayjs/ws@3.10.6 latest
+npm dist-tag add @midwayjs/ws@3.10.7 beta
+tnpm dist-tag add @midwayjs/fc-starter@3.10.6 latest
+tnpm dist-tag add @midwayjs/fc-starter@3.10.7 beta
+tnpm dist-tag add @midwayjs/async-hooks-context-manager@3.10.6 latest
+tnpm dist-tag add @midwayjs/async-hooks-context-manager@3.10.7 beta
+tnpm dist-tag add @midwayjs/axios@3.10.6 latest
+tnpm dist-tag add @midwayjs/axios@3.10.7 beta
+tnpm dist-tag add @midwayjs/bootstrap@3.10.6 latest
+tnpm dist-tag add @midwayjs/bootstrap@3.10.7 beta
+tnpm dist-tag add @midwayjs/bull-board@3.10.6 latest
+tnpm dist-tag add @midwayjs/bull-board@3.10.7 beta
+tnpm dist-tag add @midwayjs/bull@3.10.6 latest
+tnpm dist-tag add @midwayjs/bull@3.10.7 beta
+tnpm dist-tag add @midwayjs/cache@3.10.6 latest
+tnpm dist-tag add @midwayjs/cache@3.10.7 beta
+tnpm dist-tag add @midwayjs/captcha@3.10.6 latest
+tnpm dist-tag add @midwayjs/captcha@3.10.7 beta
+tnpm dist-tag add @midwayjs/casbin-redis-adapter@3.10.6 latest
+tnpm dist-tag add @midwayjs/casbin-redis-adapter@3.10.7 beta
+tnpm dist-tag add @midwayjs/casbin-typeorm-adapter@3.10.6 latest
+tnpm dist-tag add @midwayjs/casbin-typeorm-adapter@3.10.7 beta
+tnpm dist-tag add @midwayjs/casbin@3.10.6 latest
+tnpm dist-tag add @midwayjs/casbin@3.10.7 beta
+tnpm dist-tag add @midwayjs/code-dye@3.10.6 latest
+tnpm dist-tag add @midwayjs/code-dye@3.10.7 beta
+tnpm dist-tag add @midwayjs/consul@3.10.6 latest
+tnpm dist-tag add @midwayjs/consul@3.10.7 beta
+tnpm dist-tag add @midwayjs/core@3.10.6 latest
+tnpm dist-tag add @midwayjs/core@3.10.7 beta
+tnpm dist-tag add @midwayjs/cos@3.10.6 latest
+tnpm dist-tag add @midwayjs/cos@3.10.7 beta
+tnpm dist-tag add @midwayjs/cross-domain@3.10.6 latest
+tnpm dist-tag add @midwayjs/cross-domain@3.10.7 beta
+tnpm dist-tag add @midwayjs/decorator@3.10.6 latest
+tnpm dist-tag add @midwayjs/decorator@3.10.7 beta
+tnpm dist-tag add @midwayjs/etcd@3.10.6 latest
+tnpm dist-tag add @midwayjs/etcd@3.10.7 beta
+tnpm dist-tag add @midwayjs/express-session@3.10.6 latest
+tnpm dist-tag add @midwayjs/express-session@3.10.7 beta
+tnpm dist-tag add @midwayjs/faas@3.10.6 latest
+tnpm dist-tag add @midwayjs/faas@3.10.7 beta
+tnpm dist-tag add @midwayjs/grpc@3.10.6 latest
+tnpm dist-tag add @midwayjs/grpc@3.10.7 beta
+tnpm dist-tag add @midwayjs/http-proxy@3.10.6 latest
+tnpm dist-tag add @midwayjs/http-proxy@3.10.7 beta
+tnpm dist-tag add @midwayjs/i18n@3.10.6 latest
+tnpm dist-tag add @midwayjs/i18n@3.10.7 beta
+tnpm dist-tag add @midwayjs/info@3.10.6 latest
+tnpm dist-tag add @midwayjs/info@3.10.7 beta
+tnpm dist-tag add @midwayjs/jwt@3.10.6 latest
+tnpm dist-tag add @midwayjs/jwt@3.10.7 beta
+tnpm dist-tag add @midwayjs/kafka@3.10.6 latest
+tnpm dist-tag add @midwayjs/kafka@3.10.7 beta
+tnpm dist-tag add @midwayjs/mikro@3.10.6 latest
+tnpm dist-tag add @midwayjs/mikro@3.10.7 beta
+tnpm dist-tag add @midwayjs/mock@3.10.6 latest
+tnpm dist-tag add @midwayjs/mock@3.10.7 beta
+tnpm dist-tag add @midwayjs/mongoose@3.10.6 latest
+tnpm dist-tag add @midwayjs/mongoose@3.10.7 beta
+tnpm dist-tag add @midwayjs/oss@3.10.6 latest
+tnpm dist-tag add @midwayjs/oss@3.10.7 beta
+tnpm dist-tag add @midwayjs/otel@3.10.6 latest
+tnpm dist-tag add @midwayjs/otel@3.10.7 beta
+tnpm dist-tag add @midwayjs/passport@3.10.6 latest
+tnpm dist-tag add @midwayjs/passport@3.10.7 beta
+tnpm dist-tag add @midwayjs/process-agent@3.10.6 latest
+tnpm dist-tag add @midwayjs/process-agent@3.10.7 beta
+tnpm dist-tag add @midwayjs/prometheus-socket-io@3.10.6 latest
+tnpm dist-tag add @midwayjs/prometheus-socket-io@3.10.7 beta
+tnpm dist-tag add @midwayjs/prometheus@3.10.6 latest
+tnpm dist-tag add @midwayjs/prometheus@3.10.7 beta
+tnpm dist-tag add @midwayjs/rabbitmq@3.10.6 latest
+tnpm dist-tag add @midwayjs/rabbitmq@3.10.7 beta
+tnpm dist-tag add @midwayjs/redis@3.10.6 latest
+tnpm dist-tag add @midwayjs/redis@3.10.7 beta
+tnpm dist-tag add @midwayjs/security@3.10.6 latest
+tnpm dist-tag add @midwayjs/security@3.10.7 beta
+tnpm dist-tag add @midwayjs/sequelize@3.10.6 latest
+tnpm dist-tag add @midwayjs/sequelize@3.10.7 beta
+tnpm dist-tag add @midwayjs/session@3.10.6 latest
+tnpm dist-tag add @midwayjs/session@3.10.7 beta
+tnpm dist-tag add @midwayjs/socketio@3.10.6 latest
+tnpm dist-tag add @midwayjs/socketio@3.10.7 beta
+tnpm dist-tag add @midwayjs/static-file@3.10.6 latest
+tnpm dist-tag add @midwayjs/static-file@3.10.7 beta
+tnpm dist-tag add @midwayjs/swagger@3.10.6 latest
+tnpm dist-tag add @midwayjs/swagger@3.10.7 beta
+tnpm dist-tag add @midwayjs/tablestore@3.10.6 latest
+tnpm dist-tag add @midwayjs/tablestore@3.10.7 beta
+tnpm dist-tag add @midwayjs/typegoose@3.10.6 latest
+tnpm dist-tag add @midwayjs/typegoose@3.10.7 beta
+tnpm dist-tag add @midwayjs/typeorm@3.10.6 latest
+tnpm dist-tag add @midwayjs/typeorm@3.10.7 beta
+tnpm dist-tag add @midwayjs/upload@3.10.6 latest
+tnpm dist-tag add @midwayjs/upload@3.10.7 beta
+tnpm dist-tag add @midwayjs/validate@3.10.6 latest
+tnpm dist-tag add @midwayjs/validate@3.10.7 beta
+tnpm dist-tag add @midwayjs/version@3.10.6 latest
+tnpm dist-tag add @midwayjs/version@3.10.7 beta
+tnpm dist-tag add @midwayjs/view-ejs@3.10.6 latest
+tnpm dist-tag add @midwayjs/view-ejs@3.10.7 beta
+tnpm dist-tag add @midwayjs/view-nunjucks@3.10.6 latest
+tnpm dist-tag add @midwayjs/view-nunjucks@3.10.7 beta
+tnpm dist-tag add @midwayjs/view@3.10.6 latest
+tnpm dist-tag add @midwayjs/view@3.10.7 beta
+tnpm dist-tag add @midwayjs/express@3.10.6 latest
+tnpm dist-tag add @midwayjs/express@3.10.7 beta
+tnpm dist-tag add @midwayjs/koa@3.10.6 latest
+tnpm dist-tag add @midwayjs/koa@3.10.7 beta
+tnpm dist-tag add @midwayjs/web@3.10.6 latest
+tnpm dist-tag add @midwayjs/web@3.10.7 beta
+tnpm dist-tag add @midwayjs/ws@3.10.6 latest
+tnpm dist-tag add @midwayjs/ws@3.10.7 beta
+
+# Changes:
+
+#  - @midwayjs/fc-starter: 3.10.6 => 3.10.7
+#  - @midwayjs/async-hooks-context-manager: 3.10.6 => 3.10.7
+#  - @midwayjs/axios: 3.10.6 => 3.10.7
+#  - @midwayjs/bootstrap: 3.10.6 => 3.10.7
+#  - @midwayjs/bull-board: 3.10.6 => 3.10.7
+#  - @midwayjs/bull: 3.10.6 => 3.10.7
+#  - @midwayjs/cache: 3.10.6 => 3.10.7
+#  - @midwayjs/captcha: 3.10.6 => 3.10.7
+#  - @midwayjs/casbin-redis-adapter: 3.10.6 => 3.10.7
+#  - @midwayjs/casbin-typeorm-adapter: 3.10.6 => 3.10.7
+#  - @midwayjs/casbin: 3.10.6 => 3.10.7
+#  - @midwayjs/code-dye: 3.10.6 => 3.10.7
+#  - @midwayjs/consul: 3.10.6 => 3.10.7
+#  - @midwayjs/core: 3.10.6 => 3.10.7
+#  - @midwayjs/cos: 3.10.6 => 3.10.7
+#  - @midwayjs/cross-domain: 3.10.6 => 3.10.7
+#  - @midwayjs/decorator: 3.10.6 => 3.10.7
+#  - @midwayjs/etcd: 3.10.6 => 3.10.7
+#  - @midwayjs/express-session: 3.10.6 => 3.10.7
+#  - @midwayjs/faas: 3.10.6 => 3.10.7
+#  - @midwayjs/grpc: 3.10.6 => 3.10.7
+#  - @midwayjs/http-proxy: 3.10.6 => 3.10.7
+#  - @midwayjs/i18n: 3.10.6 => 3.10.7
+#  - @midwayjs/info: 3.10.6 => 3.10.7
+#  - @midwayjs/jwt: 3.10.6 => 3.10.7
+#  - @midwayjs/kafka: 3.10.6 => 3.10.7
+#  - @midwayjs/mikro: 3.10.6 => 3.10.7
+#  - @midwayjs/mock: 3.10.6 => 3.10.7
+#  - @midwayjs/mongoose: 3.10.6 => 3.10.7
+#  - @midwayjs/oss: 3.10.6 => 3.10.7
+#  - @midwayjs/otel: 3.10.6 => 3.10.7
+#  - @midwayjs/passport: 3.10.6 => 3.10.7
+#  - @midwayjs/process-agent: 3.10.6 => 3.10.7
+#  - @midwayjs/prometheus-socket-io: 3.10.6 => 3.10.7
+#  - @midwayjs/prometheus: 3.10.6 => 3.10.7
+#  - @midwayjs/rabbitmq: 3.10.6 => 3.10.7
+#  - @midwayjs/redis: 3.10.6 => 3.10.7
+#  - @midwayjs/security: 3.10.6 => 3.10.7
+#  - @midwayjs/sequelize: 3.10.6 => 3.10.7
+#  - @midwayjs/session: 3.10.6 => 3.10.7
+#  - @midwayjs/socketio: 3.10.6 => 3.10.7
+#  - @midwayjs/static-file: 3.10.6 => 3.10.7
+#  - @midwayjs/swagger: 3.10.6 => 3.10.7
+#  - @midwayjs/tablestore: 3.10.6 => 3.10.7
+#  - @midwayjs/typegoose: 3.10.6 => 3.10.7
+#  - @midwayjs/typeorm: 3.10.6 => 3.10.7
+#  - @midwayjs/upload: 3.10.6 => 3.10.7
+#  - @midwayjs/validate: 3.10.6 => 3.10.7
+#  - @midwayjs/version: 3.10.6 => 3.10.7
+#  - @midwayjs/view-ejs: 3.10.6 => 3.10.7
+#  - @midwayjs/view-nunjucks: 3.10.6 => 3.10.7
+#  - @midwayjs/view: 3.10.6 => 3.10.7
+#  - @midwayjs/express: 3.10.6 => 3.10.7
+#  - @midwayjs/koa: 3.10.6 => 3.10.7
+#  - @midwayjs/web: 3.10.6 => 3.10.7
+#  - @midwayjs/ws: 3.10.6 => 3.10.7